new gui, new function

+++ ServerSaveFiles/src/main/java/ServerPack/HandlerSerializable.java
@@ -1,17 +1,17 @@
 package ServerPack;
 
-
 import BaseData.SqLite;
+import Friend.*;
 import Message.*;
 import io.netty.channel.ChannelHandlerContext;
 import io.netty.channel.SimpleChannelInboundHandler;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-
 import java.io.*;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
+import java.nio.file.*;
+import java.nio.file.attribute.BasicFileAttributes;
+import java.util.ArrayList;
+import java.util.HashSet;
 
 public class HandlerSerializable extends SimpleChannelInboundHandler <Message> {
     private static final Logger LOGGER = LogManager.getLogger(HandlerSerializable.class);
@@ -20,6 +20,11 @@ public class HandlerSerializable extends SimpleChannelInboundHandler <Message> {
     private ClickOperation clickOperation;
     private static final int SIZE_BUFFER = 100000;
     private byte[] buffer = new byte[SIZE_BUFFER];
+    private Server server;
+
+    HandlerSerializable(Server server) {
+        this.server = server;
+    }
 
     @Override
     public void channelActive(ChannelHandlerContext ctx) throws Exception {
@@ -36,9 +41,13 @@ public class HandlerSerializable extends SimpleChannelInboundHandler <Message> {
                 LOGGER.info("Client is authorisation");
                 SqLite.disConnect();
                 dirServer = Paths.get("ServerSaveFiles/ServerClientDirectory/" + dir);
+                server.getListClient().add(new HandlerClient(auth.getNick(), dirServer, channelHandlerContext));
                 dirServerPath = dirServer.toString();
                 channelHandlerContext.writeAndFlush(new AuthMessage(true));
                 channelHandlerContext.writeAndFlush(new ListServer(clickOperation.getArrays(dirServer.toFile()), dirServer.toString()));
+                for (HandlerClient c : server.getListClient()) {
+                    c.getChannel().writeAndFlush(new ListClientMessage(getListClient()));
+                }
             } else {
                 channelHandlerContext.writeAndFlush(new AuthMessage(false));
                 LOGGER.info("Client is not authorisation");
@@ -69,14 +78,14 @@ public class HandlerSerializable extends SimpleChannelInboundHandler <Message> {
             File file = new File(nameFile);
             if (file.isDirectory()) {
                 channelHandlerContext.writeAndFlush(new ListServer(clickOperation.getArrays(file), file.getPath()));
-                dirServer = Paths.get(dirServer.toString() + nameFile);
+                dirServer = Paths.get(dirServer.toString() + "/" + file.getName());
             }
         }
 
         if (message instanceof BackMessage) {
             BackMessage back = (BackMessage) message;
             File childFile = new File(back.getChildPath());
-            if (!(childFile.getPath().equals(dirServerPath))){
+            if (!(childFile.getPath().equals(dirServerPath))) {
                 dirServer = Paths.get(childFile.getParentFile().getPath());
                 channelHandlerContext.writeAndFlush(new ListServer(clickOperation.getArrays(
                         dirServer.toFile()), dirServer.toString()));
@@ -86,11 +95,13 @@ public class HandlerSerializable extends SimpleChannelInboundHandler <Message> {
         if (message instanceof RenameMessage) {
             RenameMessage rename = (RenameMessage) message;
             File oldFile = new File(rename.getOldPath());
-            File newFile = new File(oldFile.getParentFile().toString() +"/" + rename.getNewPath());
-            oldFile.renameTo(newFile);
-            dirServer = Paths.get(newFile.getParentFile().getPath());
-            channelHandlerContext.writeAndFlush(new ListServer(clickOperation.getArrays(dirServer.toFile()), dirServer.toString()));
-            LOGGER.info("File " + rename.getOldPath() + " rename " + newFile.getName()) ;
+            if (oldFile.isFile()) {
+                File newFile = new File(oldFile.getParentFile().toString() + "/" + rename.getNewPath());
+                oldFile.renameTo(newFile);
+                dirServer = Paths.get(newFile.getParentFile().getPath());
+                channelHandlerContext.writeAndFlush(new ListServer(clickOperation.getArrays(dirServer.toFile()), dirServer.toString()));
+                LOGGER.info("File " + rename.getOldPath() + " rename " + newFile.getName());
+            }
         }
 
         if (message instanceof DeleteMessage) {
@@ -103,9 +114,10 @@ public class HandlerSerializable extends SimpleChannelInboundHandler <Message> {
                 LOGGER.info("file " + deleteFile.getLineOne() + " is delete");
             }
         }
+
         if (message instanceof SendMessage) {
             SendMessage sendMessage = (SendMessage) message;
-            try (OutputStream out = new FileOutputStream("ServerSaveFiles\\ServerClientDirectory\\newFile\\"
+            try (OutputStream out = new FileOutputStream(dirServer.toAbsolutePath() + "\\"
                     + sendMessage.getSendFile(), true)) {
                 out.write(sendMessage.getBuffer());
                 channelHandlerContext.writeAndFlush(new ListServer(clickOperation.getArrays(dirServer.toFile()), dirServer.toString()));
@@ -116,18 +128,114 @@ public class HandlerSerializable extends SimpleChannelInboundHandler <Message> {
         if (message instanceof DownloadMessage) {
             DownloadMessage downloadMessage = (DownloadMessage) message;
             File downLoadFile = new File(downloadMessage.getFileNameDownload());
-            try (InputStream in = new FileInputStream(downLoadFile)) {
-                int count = (int) (downLoadFile.length() - 1)/ SIZE_BUFFER + 1;
-                for(int i = 0; i < count; i++) {
-                    in.read(buffer);
-                    channelHandlerContext.writeAndFlush(new DownloadMessage(buffer, downLoadFile.getName()));
+            if (downLoadFile.isFile()) {
+                try (InputStream in = new FileInputStream(downLoadFile)) {
+                    int count = (int) (downLoadFile.length() - 1) / SIZE_BUFFER + 1;
+                    for (int i = 0; i < count; i++) {
+                        int read = in.read(buffer);
+                        if (read < SIZE_BUFFER) {
+                            byte[] tmp = new byte[read];
+                            System.arraycopy(buffer, 0, tmp, 0, read);
+                            channelHandlerContext.writeAndFlush(new DownloadMessage(tmp, downLoadFile.getName()));
+                        } else {
+                            channelHandlerContext.writeAndFlush(new DownloadMessage(buffer, downLoadFile.getName()));
+                        }
+                    }
+                }
+                LOGGER.info(downLoadFile.getName() + " is send to client");
+            }
+        }
+
+        if (message instanceof FriendAddMessage) {
+            FriendAddMessage mess = (FriendAddMessage) message;
+            ChannelHandlerContext channel = getChannel(mess.getNickFriend());
+            channel.writeAndFlush(message);
+        }
+
+        if (message instanceof FriendConsent) {
+            FriendConsent consent = (FriendConsent) message;
+            ChannelHandlerContext channel = getChannel(consent.getNickFriend());
+            channel.writeAndFlush(message);
+        }
+
+        if (message instanceof ShowDirFriend) {
+            ShowDirFriend dir = (ShowDirFriend) message;
+            Path dirFriend = getPath(dir.getNick());
+            File file = new File(dirFriend.toString());
+            channelHandlerContext.writeAndFlush(new ShowDirFriend(clickOperation.getArrays(file)));
+        }
+
+        if (message instanceof DownloadFileFriendInfo) {
+            DownloadFileFriendInfo friend = (DownloadFileFriendInfo) message;
+            for (HandlerClient c : server.getListClient()) {
+                if (friend.getNickFriend().equals(c.getNick())) {
+                    Files.walkFileTree(c.getClientDir(), new HashSet<>(), 4, new SimpleFileVisitor<Path>() {
+                        @Override
+                        public FileVisitResult visitFile(Path file, BasicFileAttributes attrs) throws IOException {
+                            if (friend.getNameFile().equals(file.getFileName().toString())) {
+                                File downloadFile = new File(file.toAbsolutePath().toString());
+                                channelHandlerContext.writeAndFlush(new DownloadFileFriendInfo(downloadFile.getName(), downloadFile.length()));
+                                try (InputStream inputStream = new FileInputStream(downloadFile)) {
+                                    int read = 0;
+                                    while ((read = inputStream.read(buffer)) != -1) {
+                                        if (read < SIZE_BUFFER) {
+                                            byte[] tmp = new byte[read];
+                                            System.arraycopy(buffer, 0, tmp, 0, read);
+                                            channelHandlerContext.writeAndFlush(new Buffer(tmp));
+                                        } else {
+                                            channelHandlerContext.writeAndFlush(new Buffer(buffer));
+                                        }
+                                    }
+                                }
+                                return super.visitFile(file, attrs);
+                            }
+
+                            return super.visitFile(file, attrs);
+                        }
+                    });
                 }
             }
-            LOGGER.info(downLoadFile.getName() + " is send to client");
+        }
+
+        if (message instanceof ClickFriend) {
+            ClickFriend click = (ClickFriend) message;
+            Files.walkFileTree(getPath(click.getNickFriend()), new SimpleFileVisitor<Path>() {
+                @Override
+                public FileVisitResult preVisitDirectory(Path file, BasicFileAttributes attrs) throws IOException {
+                    if (click.getFileDirName().equals(file.getFileName().toString())) {
+                        File childDir = new File(file.toString());
+                        channelHandlerContext.writeAndFlush(new ShowDirFriend(clickOperation.getArrays(childDir)));
+                    }
+                    return super.visitFile(file, attrs);
+                }
+            });
         }
     }
+
     @Override
     public void exceptionCaught(ChannelHandlerContext ctx, Throwable cause) throws Exception {
            LOGGER.info(cause.getMessage());
+           server.getListClient().removeIf(c -> ctx.equals(c.getChannel()));
+           for(HandlerClient c : server.getListClient()) {
+               c.getChannel().writeAndFlush(new ListClientMessage(getListClient()));
+           }
+    }
+
+    public ArrayList<String> getListClient() {
+       ArrayList<String> list = new ArrayList<>();
+       for(HandlerClient c : server.getListClient()) {
+           list.add(c.getNick());
+       }
+        return list;
+    }
+
+    public ChannelHandlerContext getChannel(String nick) {
+        return server.getListClient().stream().filter(c -> nick.equals(c.getNick()))
+                .findFirst().map(HandlerClient::getChannel).orElse(null);
+    }
+
+    public Path getPath(String nick) {
+       return server.getListClient().stream().filter(c -> nick.equals(c.getNick()))
+                .findFirst().map(HandlerClient::getClientDir).orElse(null);
     }
 }
